requirements.txt
requests
discord.py
tabulate
